+++ SoundLib/Audio/OggAudio.h
@@ -24,7 +24,7 @@ public:
 	OggAudio& operator=(OggAudio&&) = default;
 
 	bool Load(TString filePath);
-	long Read(BYTE* pBuffer, DWORD bufSize);
+	long Read(BYTE* pBuffer, long bufSize);
 	void Reset();
 
 private:
