+++ SoundLib/Audio/WaveAudio.h
@@ -23,7 +23,7 @@ public:
 	WaveAudio& operator=(WaveAudio&&) = default;
 
 	bool Load(TString filePath);
-	long Read(BYTE* pBuffer, DWORD bufSize);
+	long Read(BYTE* pBuffer, long bufSize);
 	void Reset();
 
 private:
